revise paths for MacPorts.  Add file existance test in bu script

+++ bu_ft_program.sh
@@ -1,6 +1,11 @@
 #! /bin/sh
 
-rm ft_program_bu.sql
+if [ -f ft_program_bu.sql ]
+then
+    rm ft_program_bu.sql
+else
+    echo "file not found - no need to delete"
+fi
 
 if [ -f /home/kwh/hg/ft_program/test_card.pl ]; then
   echo "On EeePC"
@@ -8,7 +13,7 @@ if [ -f /home/kwh/hg/ft_program/test_card.pl ]; then
   /usr/bin/mysqldump -u ft -pft --order-by-primary --skip-extended-insert ft_program > ft_program_bu.sql
 else
   echo "Not on EeePC"
-  /sw/bin/mysqldump -u ft -pft --order-by-primary --skip-extended-insert ft_program_multi > ft_program_multiag_bu.sql
-  # /sw/bin/mysqldump -h 127.0.0.1 -u ft -pft --order-by-primary --skip-extended-insert ft_program_multi > ft_program_mulit_bu.sql
+  /opt/local/lib/mariadb/bin/mysqldump -u ft -pft --order-by-primary --skip-extended-insert ft_program_multi > ft_program_multi_bu.sql
+  # /opt/local/lib/mariadb/bin/mysqldump -h 127.0.0.1 -u ft -pft --order-by-primary --skip-extended-insert ft_program_multi > ft_program_mulit_bu.sql
 fi
